requirements.txt
vulners
colorama
pyyaml
vulners